Extract describe and test block modules
