+++ apps/language_server/lib/language_server/providers/code_lens/test/test_block.ex
@@ -1,0 +1,6 @@
+defmodule ElixirLS.LanguageServer.Providers.CodeLens.Test.TestBlock do
+  @struct_keys [:name, :describe, :line]
+
+  @enforce_keys @struct_keys
+  defstruct @struct_keys
+end
